ApiRestService refactor

+++ src/main/java/org/saebio/backup/BackupService.java
@@ -8,29 +8,21 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.time.Instant;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.stream.Collectors;
 
 public class BackupService {
-    public static void createBackup(String fileName) {
-        String databaseRoute = SampleService.getDatabaseRoute();
-        File source = new File (databaseRoute + SampleService.getDatabaseFileName());
-        File destination = new File (databaseRoute + fileName + ".db");
-        try {
-            Files.copy(source.toPath(), destination.toPath(), StandardCopyOption.REPLACE_EXISTING);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
     public static Collection<Backup> getBackups() {
         String route = SampleService.getDatabaseRoute();
         File directory = new File(route);
         String[] suffixFileFilter = new String[] {"db"};
         Collection<File> fileList = FileUtils.listFiles(directory, suffixFileFilter, true);
-        return fileList.stream()
+        Collection<Backup> backupList = fileList.stream()
                 .map(Backup::new)
                 .collect(Collectors.toList());
+        return sort(backupList);
     }
 
     public static boolean changeDatabaseToBackup(Backup backup) {
@@ -47,6 +39,7 @@ public class BackupService {
     }
 
     public static boolean removeBackup(Backup backup) {
+        if (backup.selected) return false;
         File backupFile = new File(SampleService.getDatabaseRoute() + SampleService.getBackupsRoute() + backup.getName());
         return FileUtils.deleteQuietly(backupFile);
     }
@@ -54,5 +47,23 @@ public class BackupService {
     public static boolean backupExists(Backup backup) {
         Collection<Backup> backups = getBackups();
         return (backups.stream().anyMatch(b -> b.equals(backup)));
+    }
+
+    public static boolean removeOldestBackup() {
+        Collection<Backup> backups = getBackups();
+        return removeBackup((Backup) backups.toArray()[backups.size() - 1]);
+    }
+
+    private static Collection<Backup> sort(Collection<Backup> backupList) {
+        return backupList.stream().sorted(new Comparator<Backup>() {
+            @Override
+            public int compare(Backup backup1, Backup backup2) {
+                if (backup1.selected) return -1;
+                if (backup2.selected) return 1;
+                Instant instant1 = Instant.parse(backup1.createdAt);
+                Instant instant2 = Instant.parse(backup2.createdAt);
+                return instant2.compareTo(instant1);
+            }
+        }).collect(Collectors.toList());
     }
 }
